+++ mymixtape-api/constants/api.go
@@ -1,0 +1,9 @@
+package constants
+
+const (
+	SCHEME = "https"
+	HOST = "api.spotify.com"
+	VERSION = "v1"
+
+	TOKEN_URL = "https://accounts.spotify.com/api/token"
+)
