+++ mymixtape-api/constants/error.go
@@ -1,4 +1,4 @@
-package utils
+package constants
 
 import "net/http"
