+++ mymixtape-api/constants/variables.go
@@ -1,0 +1,9 @@
+package constants
+
+const (
+	STATE_LENGTH = 16
+)
+
+var (
+	CHARACTERS = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789")
+)
