+++ mymixtape-api/internal/models.go
@@ -1,0 +1,7 @@
+package internal
+
+type PlaylistTracks struct {
+	Tracks []struct {
+		ID string `json:"id"`
+	}
+}
