+++ mymixtape-api/internal/requestmanager.go
@@ -1,4 +1,5 @@
 package internal
+
 import (
 	"bytes"
 	"encoding/json"
@@ -9,6 +10,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/mymixtape-api/constants"
 	"github.com/mymixtape-api/models"
 	"github.com/mymixtape-api/utils"
 )
@@ -71,10 +73,10 @@ func (rm *RequestManager) SetToken(code string, redirect_uri string, client_id s
 	}
 
 	if response.StatusCode < 200 || response.StatusCode > 299 {
-		err, ok := utils.StatusToError[response.StatusCode]
+		err, ok := constants.StatusToError[response.StatusCode]
 
 		if !ok {
-			err = utils.Error{
+			err = constants.Error{
 				Message: "unknown error reason",
 				StatusCode: response.StatusCode,
 			}
@@ -169,9 +171,9 @@ func (rm *RequestManager) DoRequest(method, endpoint string, body io.Reader, tok
 	}
 
 	if response.StatusCode < 200 || response.StatusCode > 299 {
-		err, ok := utils.StatusToError[response.StatusCode]
+		err, ok := constants.StatusToError[response.StatusCode]
 		if !ok {
-			err = utils.Error{
+			err = constants.Error{
 				Message: "unknown error reason",
 				StatusCode: response.StatusCode,
 			}
