+++ mymixtape-api/server.go
@@ -3,8 +3,6 @@ package main
 import "github.com/mymixtape-api/router"
 
 func main() {
-	
 	engine := router.InitRoutes()
-
 	engine.Run()
 }
