+++ mymixtape-api/utils/format.go
@@ -1,16 +1,12 @@
 package utils
 
 import (
-	b64 "encoding/base64"
 	"fmt"
+	b64 "encoding/base64"
 )
 
 func GetEncodedClientCredentials(client_id string, client_secret string) string {
 	str := fmt.Sprintf("%v:%v", client_id, client_secret)
 	encodedStr := b64.StdEncoding.EncodeToString([]byte(str))
 	return encodedStr
-}
-
-func GetSpotifyTrackURI(track_id string) string {
-	return "spotify:track:" + track_id
 }
